refactor: improvement
