+++ src/css/app.css
@@ -4,3 +4,17 @@
   box-sizing: border-box;
   list-style: none;
 }
+
+::-webkit-scrollbar-thumb {
+  background: white;
+  height: 100px;
+  width: 1px;
+}
+
+::-webkit-scrollbar {
+  width: 1px;
+}
+
+:-webkit-scrollbar-track {
+  background-color: #000;
+}
